fix: address all remaining review findings
- Pin Temporal CLI to v1.3.0 in CI workflow (was unpinned master)
- Split batch/prefix ops from storage/ops.py (715→510 lines) into
  storage/batch.py (268 lines) with re-exports for backward compat
- Add multi-key warning to DaprSecretStore.get_bulk() with docstring
