+++ application_sdk/templates/incremental_sql_metadata_extractor.py
@@ -431,7 +431,7 @@ class IncrementalSqlMetadataExtractor(SqlMetadataExtractor):
         from application_sdk.execution._temporal.activity_utils import (
             get_object_store_prefix,
         )
-        from application_sdk.storage.ops import download_prefix, upload_prefix
+        from application_sdk.storage.batch import download_prefix, upload_prefix
 
         if not input.output_path:
             raise FileNotFoundError(
